Muhammed Hamza

+++ functions/purchase_functions.py
@@ -70,7 +70,7 @@ def dump_stock_difference(self) -> bool:
 def create_serial_number(self, _dict: dict) -> bool:
     added_new_serial_number = False
     while True:
-        new_serial_number = fs.get_float("Add/Remove a Serial Number? (no)", True)
+        new_serial_number = fs.get_float(self, "Add/Remove a Serial Number? (no)", True)
         if new_serial_number:
             if new_serial_number < 0:
                 if int(new_serial_number) * -1 in _dict["serial-numbers"]:
@@ -89,13 +89,15 @@ def create_serial_number(self, _dict: dict) -> bool:
 
 
 def create_supplier(self) -> dict[str, Any]:
-    extra_str = (
-        "شركه صاله بىع لبىع تعبئه تحمىص اجود انواع مواد غذائى غذائىه شوكولاطه شوكولاته حلوى "
-        "حلوىات بقولىات توابل مكسرات تمر تمور عطر عطور تنظىف نظافه جمله قطاعى نقدا"
+    if not fs.permisstion(self, "create-supplier"):
+        fs.check_quit(self, "00", _clear_terminal=False)
+    extra_str_ar = (
+        "شركه صاله ثلاجة بىع لبىع تعبئه تحمىص اجود انواع مواد غذائى غذائىه شوكولاطه شوكولاته حلوى "
+        "حلوىات بقولىات توابل مكسرات تمر تمور لحم لحوم مجمد مجمدات حفظ تبرىد عطر عطور تنظىف نظافه جمله قطاعى نقدا"
     )
     suppliers = fs.load_data(Path("data/suppliers.json"))
     while True:
-        supplier_name = str(fs.get_str("Supplier Name"))
+        supplier_name = str(fs.get_str(self, "Supplier Name"))
         name_already_exists = False
         for supplier in suppliers:
             if fs.sanitized_and_desplited(supplier_name.lower()) == fs.sanitized_and_desplited(
@@ -112,7 +114,7 @@ def create_supplier(self) -> dict[str, Any]:
     sanitized_supplier = fs.sanitized_and_desplited(supplier_name.copy())
 
     for index, word in enumerate(sanitized_supplier.split()):
-        if word not in extra_str:
+        if word not in extra_str_ar:
             shorted_supplier_name.append(supplier_name[index])
 
     self.cache["last-supplier-number"] += 1
@@ -132,8 +134,10 @@ def get_supplier(self) -> dict[str, Any]:
     suppliers = fs.load_data(Path("data/suppliers.json"))
     if not suppliers:
         return create_supplier(self)
+    _permisstion = fs.permisstion(self, "create-supplier", False)
+    message = "Supplier Name/Number" + " (new supplier)" if _permisstion else str()
     while True:
-        str_or_float = fs.get_str_or_float("Supplier Name/Number (new supplier)", True)
+        str_or_float = fs.get_str_or_float(self, message, _permisstion)
         if not str_or_float:
             return create_supplier(self)
         if type(str_or_float) is str:
@@ -146,8 +150,10 @@ def get_supplier(self) -> dict[str, Any]:
 
 
 def create_item(self, serial_number: int) -> dict[str, Any]:
+    if not fs.permisstion(self, "create-item"):
+        fs.check_quit(self, "00", _clear_terminal=False)
     while True:
-        name = fs.get_str("Item Name")
+        name = fs.get_str(self, "Item Name")
         name_already_exists = False
         for item in self.items:
             if fs.sanitized_and_desplited(name.lower()) == fs.sanitized_and_desplited(item["item-name"].lower()):
@@ -156,28 +162,28 @@ def create_item(self, serial_number: int) -> dict[str, Any]:
                 break
         if not name_already_exists:
             break
-    pack = fs.get_float("Pack (no pack)", True)
+    pack = fs.get_float(self, "Pack (no pack)", True)
     imported_packs = None
     if pack is not None:
-        imported_packs = float(fs.get_float("Imported Packs"))
+        imported_packs = float(fs.get_float(self, "Imported Packs"))
         imported = imported_packs * pack
     else:
-        imported = float(fs.get_float("Imported Items"))
-    in_stock = fs.get_float(f"In-Stock Items ({imported})", True)
+        imported = float(fs.get_float(self, "Imported Items"))
+    in_stock = fs.get_float(self, f"In-Stock Items ({imported})", True)
     in_stock = in_stock if in_stock is not None else imported
     purchase_pack_price = None
     if pack is not None:
-        purchase_pack_price = float(fs.get_float("Purchase pack Price"))
+        purchase_pack_price = float(fs.get_float(self, "Purchase pack Price"))
         purchase_price = purchase_pack_price / pack
     else:  # self.settings.profit, self.settings.divide
-        purchase_price = float(fs.get_float("Purchase Price"))
+        purchase_price = float(fs.get_float(self, "Purchase Price"))
     profit = float(self.calculate_profit(purchase_price))
-    sell_price = fs.get_float(f"Sell Price ({purchase_price} -> {profit})", True)
+    sell_price = fs.get_float(self, f"Sell Price ({purchase_price} -> {profit})", True)
     sell_price = sell_price if sell_price is not None else profit
     sell_pack_price = None
     if purchase_pack_price:
         pack_profit = float(self.calculate_profit(purchase_pack_price))
-        sell_pack_price = fs.get_float(f"Sell Pack Price ({purchase_pack_price} -> {pack_profit})", True)
+        sell_pack_price = fs.get_float(self, f"Sell Pack Price ({purchase_pack_price} -> {pack_profit})", True)
         sell_pack_price = sell_pack_price if sell_pack_price is not None else pack_profit
     item = {
         "serial-numbers": [serial_number],
@@ -200,7 +206,9 @@ def create_item(self, serial_number: int) -> dict[str, Any]:
 
 def get_item(self) -> tuple[dict | None, str]:
     serial_number = int()
-    str_or_float = fs.get_str_or_float("Item Name/Serial Number (*)", True)
+    _permisstion = fs.permisstion(self, "create-purchase-item", False)
+    message = "Item Name/Serial Number" + " (*)" if _permisstion else str()
+    str_or_float = fs.get_str_or_float(self, message, _permisstion)
     if not str_or_float:
         self.cache["last-barcode-number"] += 1
         return create_item(self, self.cache["last-barcode-number"]), "create"
@@ -219,39 +227,43 @@ def get_item(self) -> tuple[dict | None, str]:
         serial_number = int(str_or_float)
     if type(found_items) is list and len(found_items) == 1:
         message = f"Are you sure you want to REMOVE item: {found_items[0]['item-name']}? (Yes)"
-        if is_delete and not fs.get_str(message, True):
+        if is_delete and fs.permisstion(self, "delete-purchase-item") and not fs.get_str(self, message, True):
             return delete_item(self, found_items[0]), str()
         return found_items[0], "get"
-    if serial_number:
+    if serial_number and _permisstion:
         return create_item(self, serial_number), "create"
     return None, str()
 
 
 def update_item(self, item: dict) -> dict[str, Any]:
+    if not fs.permisstion(self, "update-item"):
+        fs.check_quit(self, "00", _clear_terminal=False)
     pack = item["pack"]
     imported_packs = None
     if pack is not None:
-        imported_packs = fs.get_float("Imported Packs")
+        imported_packs = fs.get_float(self, "Imported Packs")
         imported = imported_packs * pack
     else:
-        imported = fs.get_float("Imported Items")
-    in_stock = fs.get_float(f"In-Stock Items ({imported})", True) or imported
+        imported = fs.get_float(self, "Imported Items")
+    in_stock = fs.get_float(self, f"In-Stock Items ({imported})", True) or imported
     purchase_pack_price = None
     if pack is not None:
         purchase_pack_price = (
-            fs.get_float(f"Purchase pack Price ({item["purchase-pack-price"]})", True) or item["purchase-pack-price"]
+            fs.get_float(self, f"Purchase pack Price ({item["purchase-pack-price"]})", True) or item["purchase-pack-price"]
         )
         purchase_price = purchase_pack_price / pack
     else:
-        purchase_price = fs.get_float(f"Purchase Price ({item["purchase-price"]})", True) or item["purchase-price"]
+        purchase_price = fs.get_float(self, f"Purchase Price ({item["purchase-price"]})", True) or item["purchase-price"]
     if purchase_price != item["purchase-price"]:
         profit = self.calculate_profit(purchase_price)
-        sell_price = fs.get_float(f"Sell Price ({purchase_price} -> {profit})", True) or profit
+        sell_price = fs.get_float(self, f"Sell Price ({purchase_price} -> {profit})", True) or profit
         pack_profit = self.calculate_profit(purchase_pack_price)
-        sell_pack_price = fs.get_float(f"Sell Pack Price ({purchase_pack_price} -> {pack_profit})", True) or pack_profit
+        sell_pack_price = (
+            fs.get_float(self, f"Sell Pack Price ({purchase_pack_price} -> {pack_profit})", True) or pack_profit
+        )
     else:
-        sell_price = fs.get_float(f"Sell Price ({item["sell-price"]})", True) or item["sell-price"]
-        sell_pack_price = fs.get_float(f"Sell Pack Price ({item["sell-pack-price"]})", True) or item["sell-pack-price"]
+        sell_price = fs.get_float(self, f"Sell Price ({item["sell-price"]})", True) or item["sell-price"]
+        sell_pack_price = fs.get_float(self, f"Sell Pack Price ({item["sell-pack-price"]})", True) or item["sell-pack-price"]
     _item = {
         "serial-numbers": item["serial-numbers"],
         "item-name": item["item-name"],
@@ -292,11 +304,11 @@ def add_to_item(self, item: dict) -> dict[str, Any]:
     pack = item["pack"]
     imported_packs = None
     if pack is not None:
-        imported_packs = fs.get_float("Imported Packs")
+        imported_packs = fs.get_float(self, "Imported Packs")
         imported = imported_packs * pack
     else:
-        imported = fs.get_float("Imported Items")
-    in_stock = (fs.get_float(f"In-Stock Items ({imported})", True) or imported) + item["in-stock"]
+        imported = fs.get_float(self, "Imported Items")
+    in_stock = (fs.get_float(self, f"In-Stock Items ({imported})", True) or imported) + item["in-stock"]
     if imported_packs:
         imported_packs += item["imported-packs"]
     imported += item["imported-items"]
@@ -328,17 +340,20 @@ def add_to_item(self, item: dict) -> dict[str, Any]:
 
 
 def create_invoice(self) -> dict[str, Any]:
-    supplier_invoice_number = fs.get_float("Supplier Invoice Number (no number)", True)
+    if not fs.permisstion(self, "create-purchase-invoice"):
+        fs.check_quit(self, "00", _clear_terminal=False)
+    supplier_invoice_number = fs.get_float(self, "Supplier Invoice Number (no number)", True)
     supplier = get_supplier(self)
-    invoice_number = self.cache["last-inovace-number"] + 1
-    self.cache["last-inovace-number"] += 1
+    invoice_number = self.cache["last-invoice-number"] + 1
+    self.cache["last-invoice-number"] += 1
     self.cache["invoice-numbers"].append(invoice_number)
     today = datetime.date.today().isoformat()  # convert to ISO string
     return {
         "supplier": supplier,
-        "date": str(fs.validate_date(fs.get_str(f"Inovace Date ({today})", True)) or today),
+        "date": str(fs.validate_date(fs.get_str(self, f"Invoice Date ({today})", True)) or today),
         "invoice-number": invoice_number,
         "supplier-invoice-number": supplier_invoice_number,
+        "invoice-path": str(),
         "items": list(),
     }
 
@@ -346,7 +361,9 @@ def create_invoice(self) -> dict[str, Any]:
 def get_invoice(self) -> dict | None:
     if not self.cache["invoice-numbers"]:
         return create_invoice(self)
-    invoice_number = fs.get_float("Invoice Number (new invoice)", True)
+    _permisstion = fs.permisstion(self, "create-purchase-invoice", False)
+    message = "Invoice Number" + " (new invoice)" if _permisstion else str()
+    invoice_number = fs.get_float(self, message, _permisstion)
     if not invoice_number:
         return create_invoice(self)
     invoice_number = int(invoice_number)
@@ -354,21 +371,20 @@ def get_invoice(self) -> dict | None:
     if invoice_number < 0:
         invoice_number *= -1
         is_delete = True
-    invoice_path = str()
-    all_invoices = list()
-    invoices = list()
-    for invoice_path in self.purchases_path.iterdir():
-        if invoice_path.is_file():
-            invoice = fs.load_data(invoice_path)
-            all_invoices.append(invoice)
-            if invoice_number == invoice["invoice-number"]:
-                invoices.append(invoice)
+    invoices = [invoice for invoice in self.invoices if invoice_number == invoice["invoice-number"]]
     if len(invoices) == 1:
-        if is_delete and not fs.get_str(f"Are you sure you want to DELETE purchase invoice: {invoice_number}? (Yes)", True):
-            return delete_invoice(self, invoice_path, invoice_number)
+        if (
+            is_delete
+            and fs.permisstion(self, "delete-purchase-invoice")
+            and not fs.get_str(self, f"Are you sure you want to DELETE purchase invoice: {invoice_number}? (Yes)", True)
+        ):
+            return delete_invoice(self, Path(invoices[0]["invoice-path"]), invoice_number)
         return invoices[0]
     pprint(
-        [(invoice["invoice-number"], invoice["supplier"]["shorted-supplier-name"]) for invoice in (invoices or all_invoices)]
+        [
+            (invoice["invoice-number"], invoice["supplier"]["shorted-supplier-name"])
+            for invoice in (invoices or self.invoices)
+        ]
     )
 
 
@@ -448,4 +464,4 @@ def delete_invoice(self, invoice_path: Path, invoice_number: int) -> None:
     self.cache["invoice-numbers"].remove(invoice_number)
     dump_stock_difference(self)
     fs.dump_data(self.cache, Path("data/cache/purchases.json"))
-    fs.check_quit("00")
+    fs.check_quit(self, "00")
